+++ teachertool/src/transforms/loadProjectAsync.ts
@@ -1,0 +1,3 @@
+export const test = () => {
+    console.log("test");
+}
